Moved two_basic tests in its own file.
