+++ crates/lib3h/tests/test_suites/mod.rs
@@ -1,0 +1,1 @@
+pub mod two_basic;
